requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
allure-pytest==2.14.2
    # via -r requirements.in
allure-python-commons==2.14.2
    # via allure-pytest
annotated-types==0.7.0
    # via
    #   -r requirements.in
    #   pydantic
anyio==4.9.0
    # via
    #   -r requirements.in
    #   httpx
    #   openai
    #   starlette
    #   watchfiles
attrs==25.3.0
    # via allure-python-commons
certifi==2025.4.26
    # via
    #   -r requirements.in
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.2
    # via
    #   -r requirements.in
    #   requests
click==8.2.0
    # via uvicorn
distro==1.9.0
    # via
    #   -r requirements.in
    #   openai
elevenlabs==1.59.0
    # via -r requirements.in
fastapi==0.115.12
    # via -r requirements.in
h11==0.16.0
    # via
    #   -r requirements.in
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via
    #   -r requirements.in
    #   httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via
    #   -r requirements.in
    #   elevenlabs
    #   openai
idna==3.10
    # via
    #   -r requirements.in
    #   anyio
    #   httpx
    #   requests
iniconfig==2.1.0
    # via pytest
jiter==0.9.0
    # via
    #   -r requirements.in
    #   openai
openai==1.79.0
    # via -r requirements.in
packaging==25.0
    # via pytest
pluggy==1.6.0
    # via
    #   allure-python-commons
    #   pytest
pydantic==2.11.4
    # via
    #   -r requirements.in
    #   elevenlabs
    #   fastapi
    #   openai
pydantic-core==2.33.2
    # via
    #   -r requirements.in
    #   elevenlabs
    #   pydantic
pytest==8.3.5
    # via allure-pytest
python-dotenv==1.1.0
    # via uvicorn
pyyaml==6.0.2
    # via uvicorn
requests==2.32.3
    # via
    #   -r requirements.in
    #   elevenlabs
sniffio==1.3.1
    # via
    #   -r requirements.in
    #   anyio
    #   openai
starlette==0.46.2
    # via fastapi
tqdm==4.67.1
    # via
    #   -r requirements.in
    #   openai
typing-extensions==4.13.2
    # via
    #   -r requirements.in
    #   anyio
    #   elevenlabs
    #   fastapi
    #   openai
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.0
    # via
    #   -r requirements.in
    #   pydantic
urllib3==2.4.0
    # via
    #   -r requirements.in
    #   requests
uvicorn==0.34.2
    # via -r requirements.in
uvloop==0.21.0
    # via uvicorn
watchfiles==1.0.5
    # via uvicorn
websockets==15.0.1
    # via
    #   -r requirements.in
    #   elevenlabs
    #   uvicorn
